+++ spec/seo_cache_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe SeoCache do
+describe SeoCache do
   it 'has a version number' do
     expect(SeoCache::VERSION).not_to be nil
   end